refactor: move GalaxyClient
