+++ app/api/endpoints/invocation.py
@@ -16,7 +16,7 @@ from fastapi.concurrency import run_in_threadpool
 from starlette.status import HTTP_204_NO_CONTENT
 
 from app.context import current_api_key
-from app.bioblend_server.galaxy import GalaxyClient
+from app.galaxy import GalaxyClient
 from app.bioblend_server.executor.workflow_manager import WorkflowManager
 from app.api.schemas import invocation
 from app.api.socket_manager import ws_manager
