+++ app/api/endpoints/invocations/data_manager.py
@@ -7,7 +7,7 @@ from typing import Dict
 from sys import path
 path.append(".")
 
-from app.bioblend_server.galaxy import GalaxyClient
+from app.galaxy import GalaxyClient
 from bioblend.galaxy.objects.wrappers import Invocation
 from app.bioblend_server.executor.workflow_manager import WorkflowManager
 from app.orchestration.invocation_cache import InvocationCache
