+++ app/api/endpoints/invocations/invocation_service.py
@@ -8,7 +8,7 @@ path.append(".")
 from fastapi import Response
 from starlette.status import HTTP_204_NO_CONTENT
 
-from app.bioblend_server.galaxy import GalaxyClient
+from app.galaxy import GalaxyClient
 from app.bioblend_server.executor.workflow_manager import WorkflowManager
 from app.api.schemas import invocation, workflow
 from app.orchestration.invocation_cache import InvocationCache
