+++ app/bioblend_server/executor/history_manager.py
@@ -9,7 +9,7 @@ load_dotenv()
 from sys import path
 path.append('.')
 
-from app.bioblend_server.galaxy import GalaxyClient
+from app.galaxy import GalaxyClient
 from bioblend.galaxy.objects.wrappers import History
